+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
-## [1.3.0] - 2021-11-05
+## [1.3.2] - 2021-11-06
+### Added
+- The `DABreadCrumbs` class for displaying a breadcrumbs UI to the
+  user when the current question is the result of a nested action.
+
+## [1.3.1] - 2021-11-05
 ### Added
 - The `button colors` Configuration directive for customizing the
   Bootstrap colors of buttons.